refactor:  muda entidade de usuario de diretorio
